notes_service/requirements.txt
asgiref==3.6.0
Django==4.2
django-clearcache==1.2.1
django-cors-headers==4.4.0
django-debug-toolbar==4.1.0
django-environ==0.11.2
django-filter==23.1
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
gunicorn
